Use treescope for all SVG references
This patch changes how SVG references are looked up by using
the treescope instead of the document for id references. If an
SVG element is in a shadow tree, its id references will now be
confined to this tree. If an SVG element is simply in a document, its
id references will match in the document but not down into shadow trees.

This patch handles all reference lookups except for animations
(see: crbug.com/363927).

BUG=356891

Review URL: https://codereview.chromium.org/236063010

git-svn-id: svn://svn.chromium.org/blink/trunk@171797 bbb929c8-8fbe-4397-9dbb-9b2b20218538