Correct negative sub-pixel offset handling in composited layers
Code in CompositedLayerMapping used flooredIntPoint in an attempt
to represent an offset with the sub-pixel fraction removed.
Unfortunately, flooring a negative number with a sub-pixel fraction
will floor down to the next negative value, which isn't what the
code wanted. Changing it to use toInt() instead, which has the
effect of truncating the values.

BUG=362889

Review URL: https://codereview.chromium.org/240453003

git-svn-id: svn://svn.chromium.org/blink/trunk@171805 bbb929c8-8fbe-4397-9dbb-9b2b20218538