Treat flex items as selection roots.
Blink paints "selection gaps" that extend the selection to the bounds of the selection root.

If the selection spans two flex items, and the selection root is the RenderBody, the selection gap for the first item is painted on top of the second item, obscuring its text.

By making each flex item a selection root, the selection gap is painted only up to the bounds of the flex item.  This is the logic we use for table cells and positioned objects.

BUG=167341

Review URL: https://codereview.chromium.org/231793006

git-svn-id: svn://svn.chromium.org/blink/trunk@171810 bbb929c8-8fbe-4397-9dbb-9b2b20218538