Fix destruction order of Media element and MediaKeys.
When a MediaKeys object is associated with a Media element, during destruction
of the Media element, we want to destroy the media player before destroying the
CDM. This is because we assume the CDM will outlive the media pipeline.

Note that this doesn't resolve the more general issue that setMediaKeys(0) can
be called when the media pipeline is still using the CDM. That issue is tracked
by http://crbug.com/330324.

BUG=360345
TEST=Added a new layout test to cover the issue.

Review URL: https://codereview.chromium.org/238123002

git-svn-id: svn://svn.chromium.org/blink/trunk@171819 bbb929c8-8fbe-4397-9dbb-9b2b20218538