Cleanup SVGElement::isInUserAgentShadowTree and rename it inUseShadowTree
This patch renames isInUserAgentShadowTree to inUseShadowTree and
cleans up the implementation to check whether an SVGElement is in
a <use>-created shadow tree. This patch also cleans up the
remaining callers of inShadowTree() to call inUseShadowTree(), as
many of the comments suggest.

No change in behavior because we don't have SVG in user agent
shadow roots.

BUG=364302

Review URL: https://codereview.chromium.org/238293009

git-svn-id: svn://svn.chromium.org/blink/trunk@171904 bbb929c8-8fbe-4397-9dbb-9b2b20218538