Iterate Through Fewer Children When Seeking Overhanging Floats
If a container's height is reduced it doesn't need to iterate through all
its children looking for floats that might now overhang - it only needs to
look back through float-containing children and even then only as far as the last
child with a float in its list of own/intruding floats that overhangs
the container's new height.

We also don't need to track the bottom position of the lowest hanging float
we encounter; we get that for free when we look at the lowestFloatLogicalBottom()
of the last child. This is so because each child keeps a copy of any
floats intruding from the previous sibling in its float list.

Review URL: https://codereview.chromium.org/228663014

git-svn-id: svn://svn.chromium.org/blink/trunk@171906 bbb929c8-8fbe-4397-9dbb-9b2b20218538