Rework SVG sizing
The basis of this change is to map explicit width and height
attributes to CSS properties, essentially promoting them to
presentation attributes. Note that implicit "100%" width and height
are not mapped.

This enables us to remove the concept of "percentage intrinsic size"
and rely on normal CSS rules to resolve percentage values.

The change has been approved by the SVG WG and the spec is being
updated. Minutes here:

http://www.w3.org/2014/04/07-svg-minutes.html#item03

Many of the -expected updates are caused by SVG size changes in turn
affecting where line breaks are inserted.

BUG=308992

Review URL: https://codereview.chromium.org/26390004

git-svn-id: svn://svn.chromium.org/blink/trunk@171958 bbb929c8-8fbe-4397-9dbb-9b2b20218538