Incorrect invalidation on Google Forum
The issue was that we would generate incremental invalidations for
a positioned element instead of a full invalidation. This is
because the repaint rect didn't move when intersected with the
overflow clip.

The fix is to force positioned children of a positioned element
that required a positioned movement layout to be relaid out. This
is similar to what was done in
https://codereview.chromium.org/192863004. To achieve the previous
fix, we make positioned movement layout do a simplified layout,
which simplifies the code by removing the possibility of doing
both independently.

BUG=360524

Review URL: https://codereview.chromium.org/241113004

git-svn-id: svn://svn.chromium.org/blink/trunk@172042 bbb929c8-8fbe-4397-9dbb-9b2b20218538