REGRESSION(r157081): Fix multipart handling of ResourceFetcher
r157081 broke multipart handling in ResourceFetcher.
The breakage was harmless except Motion JPEG case - ResourceFetcher
lost the reference to Motion JPEG resource so it was no longer
able to cancel the ongoing resource fetching.

This change makes the tracking working again.
Also, the change adds a stopFetching() call to DocumentLoder::stopLoading()
that is used to be there but was lost at some point.

This is re-landing of r171969 with a fix.

TEST=stop-loading.html
TBR=japhet@chromium.org
BUG=309641

Review URL: https://codereview.chromium.org/245143003

git-svn-id: svn://svn.chromium.org/blink/trunk@172081 bbb929c8-8fbe-4397-9dbb-9b2b20218538