Get rid of willBeCompositedOrSquashed setting for the root layer.
As best I can tell, this doesn't have any effect when we're
processing the root layer. In order to make that more clear,
I moved a bunch of code into an rootLayer and a !rootLayer
block.

Since we always process the root layer at the top of the
recursion, we don't need to do any of the updating of the
bits for the RecursionData. Also, there's a bunch of
properties that are never set for the root layer (e.g.
blending or reflections).

Finally, we never add the root layer to the overlap map
now, so we can assert that in OverlapMap::add.

Review URL: https://codereview.chromium.org/243583011

git-svn-id: svn://svn.chromium.org/blink/trunk@172104 bbb929c8-8fbe-4397-9dbb-9b2b20218538