[Oilpan]: Get rid of the last CSSValue refptrs by using weak processing to clean up instead of destructors.
With this change we avoid the RefPtr to CSSValue in CSSCursorImageValue since we no longer need to access it from the CSSCursorImageValue's destructor. Instead we use weak members and weak callbacks to do the cleanup.

Also removed a few other RefPtrs to CSSValue that had snug in since last.

R=ager@chromium.org, erik.corry@gmail.com, haraken@chromium.org, jochen@chromium.org, oilpan-reviews@chromium.org, tkent@chromium.org, vegorov@chromium.org, zerny@chromium.org
BUG=341815

Review URL: https://codereview.chromium.org/235053002

git-svn-id: svn://svn.chromium.org/blink/trunk@172140 bbb929c8-8fbe-4397-9dbb-9b2b20218538