Remove scrollbars rom fast/images/exif-orientation-css.html
This test is larger than the viewport, but the text outside the
viewport is the console output and not part of the test's pixel
results to verify pass/fail.

Use overflow:hidden to prevent scrollbars so we don't test the
scrollbars' pixel output in this test.

R=enne@chromium.org
BUG=364946

Review URL: https://codereview.chromium.org/247793003

git-svn-id: svn://svn.chromium.org/blink/trunk@172385 bbb929c8-8fbe-4397-9dbb-9b2b20218538