Setting RL's grouped mapping should remove it from the old one.
Currently, when a RenderLayer is assigned to a new grouped
mapping for squashing, updateGroupedMapping will only mark
the old grouped mapping as needing a graphics layer update.
The old grouped mapping will, however, retain a raw pointer
to the RenderLayer.

It turns out that it is possible to get the old grouped
mapping to use this raw pointer, so it's important that we
clear it when we switch backings so that this never happens.

The attached test case triggers this use of one of these
stale pointers, but it's clunky. It should be possible to
get a better repro.

R=abarth@chromium.org
BUG=365064

Review URL: https://codereview.chromium.org/244253006

git-svn-id: svn://svn.chromium.org/blink/trunk@172409 bbb929c8-8fbe-4397-9dbb-9b2b20218538