Offset rects by origin when projecting from composited to graphics layer.
The origin for the graphics layer does not have to be the same as the
composited layer (e.g. when a child layer has negative offset and paints into
this layer), so when projecting rects to graphics layer space they have to be
offset by the origin for the composited layer.

Update internals to also return the rects with the offset set.

BUG=364346

Review URL: https://codereview.chromium.org/250193002

git-svn-id: svn://svn.chromium.org/blink/trunk@172462 bbb929c8-8fbe-4397-9dbb-9b2b20218538