Don't special case mousewheel on horizontal scrollbar on Linux
WebKit used to treat vertical wheel events on a horizontal scrollbar specially
on Linux in order to match GTK behavior.  With Linux Aura, there's no longer
any reason to treat Linux specially (we're ripping out all GTK-specific code).

Also deletes a test that was added for this when it was tweaked to work with
touch scrolls (which at the time sent wheel events).  Since theres no longer
any special code for this case, it's probably not worth testing (and it
certainly has nothing to do with touch scrolling anymore).

BUG=None

Review URL: https://codereview.chromium.org/251753004

git-svn-id: svn://svn.chromium.org/blink/trunk@172671 bbb929c8-8fbe-4397-9dbb-9b2b20218538